autorotate/requirements.txt
pdf2image==1.16.3
opencv-python==4.8.0
numpy==1.24.3
Pillow==9.5.0
fpdf==1.7.2
pytesseract=0.3.10

